+++ .env
@@ -34,5 +34,7 @@ VNP_URL = "https://sandbox.vnpayment.vn/paymentv2/vpcpay.html"
 
 VNP_ReturnUrl = "http://localhost:5000/order/sendVNPost/return"
 
+MOMO_ACCESSKEY = "F8BBA842ECF85"
 
+MOMO_SECRETKEY = "K951B6PE1waDMi640xX08PD3vg6EkVlz"
